rename : 폴더 이름 변경
- input&output 폴더를 view 폴더로 이름 수정 및 통합
- racing 폴더를 domain 폴더로 이름 수정
