+++ src/main/java/car/domain/model/CarName.java
@@ -1,4 +1,4 @@
-package car.racing.model;
+package car.domain.model;
 
 public class CarName {
