+++ src/main/java/car/domain/model/Cars.java
@@ -1,4 +1,4 @@
-package car.racing.model;
+package car.domain.model;
 
 import java.util.Arrays;
 import java.util.Collections;
