+++ src/main/java/car/view/InputService.java
@@ -1,4 +1,4 @@
-package car.input;
+package car.view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
