+++ src/test/java/car/racingtest/CarsTest.java
@@ -3,7 +3,7 @@ package car.racingtest;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import car.racing.model.Cars;
+import car.domain.model.Cars;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
